modified:   package-lock.json 	new file:   public/images/homepagetop.png 	new file:   public/images/profile_icon.png 	deleted:    public/logo192.png 	deleted:    public/logo512.png 	modified:   src/App.css 	deleted:    src/App.js 	new file:   src/App.jsx 	deleted:    src/Components/IncidentForm.js 	renamed:    src/Components/IncidentList.js -> src/Components/IncidentList.jsx 	deleted:    src/Components/Location.js 	renamed:    src/Components/LoginForm.js -> src/Components/LoginForm.jsx 	renamed:    src/Components/Logout.js -> src/Components/Logout.jsx 	renamed:    src/Components/Navbar.js -> src/Components/Navbar.jsx 	renamed:    src/Components/RegisterForm.js -> src/Components/RegisterForm.jsx 	renamed:    src/Components/UserPage.js -> src/Components/UserPage.jsx 	renamed:    src/Components/ViewReports.js -> src/Components/ViewReports.jsx 	new file:   src/Components/buttons/bigButton.jsx 	new file:   src/Components/buttons/bigButton.module.css 	new file:   src/Components/incidentPage/IncidentForm.js 	new file:   src/Components/landingpage/HomePage.jsx 	new file:   src/Components/landingpage/Homepage.module.css 	new file:   src/Components/utils/Functions/emailValidator.js 	new file:   src/Components/utils/Functions/phoneNumberValidator.js 	modified:   src/index.css
