--- src/App.js
@@ -1,31 +1,0 @@
-import React from 'react';
-import { BrowserRouter as Router, Route,Routes} from 'react-router-dom';
-import IncidentForm from './Components/IncidentForm';
-import IncidentList from './Components/IncidentList';
-import Location from './Components/Location';
-import LoginForm from './Components/LoginForm';
-import RegisterForm from './Components/RegisterForm';
-import ViewReports from './Components/ViewReports';
-import Navbar from './Components/Navbar';
-import UserPage from './Components/UserPage';
-import Logout from './Components/Logout';
-
-function App() {
-  return (
-    <Router>
-      <Navbar />
-      <Routes>
-        <Route path="/" element={<Location />} />
-        <Route path="/report-incident" element={<IncidentForm />} />
-        <Route path="/incident-list" element={<IncidentList />} />
-        <Route path="/view-reports" element={<ViewReports />} />
-        <Route path="/login" element={<LoginForm />} />
-        <Route path="/register" element={<RegisterForm />} />
-        <Route path="/user" element={<UserPage />} />
-        <Route path="/logout" element={<Logout />} />
-      </Routes>
-    </Router>
-  );
-}
-
-export default App;
