+++ src/Components/utils/Functions/emailValidator.js
@@ -1,0 +1,6 @@
+const isValidEmail = (email) => {
+    // Basic email validation
+    const emailRegex = /^[^\s@]+@[^\s@]+\.[^\s@]+$/;
+    return emailRegex.test(email);
+};
+export default isValidEmail;
